main: Color versions based on semver
This changes the output coloring of update versions based on semver. If
both versions being compared follow semver, then we have the following
cases:

- The current version is greater than or equal to the "latest" version
- The latest version has a greater major version number
- The latest version has a greater minor or patch version number

The color of the latest version is chosen based on what
`yarn update-interactive` does (sorta). In the first case above, the
color will be gray (technically light black). In the second case, red,
and lastly, yellow. This helps the user estimate at a glance what to
look for while performing and testing a package update.

If at least one of the versions cannot be parsed as semver, then the
color will simply be red.

Fixes https://github.com/AerynOS/ent/issues/9

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>